requirements.txt
flask==2.3.0
qrcode==7.4.2
pillow==10.2.0
flask-socketio==5.3.6
flask-sqlalchemy==3.1.1
pymysql==1.1.0
python-dotenv==1.0.1
cryptography==42.0.5

